Move common actions to the Abstract
Make URI comparisons case sensitive
